refactor: rename directory & file names
- overrides/ → components/
- lang-specific.ts → languages.ts
