+++ modules/eslint/src/configs/components/languages.ts
